refactor: avoid redundant naming and make layers visible
- made folders to show layers clearly
