+++ internal/repository/user/read.go
@@ -1,0 +1,1 @@
+package repository
