+++ internal/repository/user/write.go
@@ -1,0 +1,1 @@
+package repository
