--- internal/repository/user_repo/user_write_repo.go
@@ -1,1 +1,0 @@
-package Repository
