+++ internal/service/user/dto.go
@@ -1,6 +1,9 @@
-package dto
+package service
 
-import "time"
+import (
+	"anchor-blog/internal/domain/entities"
+	"time"
+)
 
 type SocialLinkDTO struct {
 	Platform string `json:"platform"`
@@ -27,4 +30,33 @@ type UserDTO struct {
 	CreatedAt time.Time      `json:"created_at"`
 	UpdatedAt time.Time      `json:"updated_at"`
 	UserPosts []string       `json:"user_posts"`
+}
+
+// :::::::::  Mapper functions  ::::::::::
+func EntityToDTO(ue entities.User) UserDTO {
+	socialLinks := make([]SocialLinkDTO, len(ue.Profile.SocialLinks))
+
+	for index, socialLink := range ue.Profile.SocialLinks {
+		socialLinks[index] = SocialLinkDTO{Platform: socialLink.Platform, URL: socialLink.URL}
+	}
+
+	return UserDTO{
+		ID:        ue.ID,
+		Username:  ue.Username,
+		FirstName: ue.FirstName,
+		LastName:  ue.LastName,
+		Email:     ue.Email,
+		Role:      ue.Role,
+		Activated: ue.Activated,
+		LastSeen:  ue.LastSeen,
+		CreatedAt: ue.CreatedAt,
+		UpdatedBy: ue.UpdatedBy,
+		UpdatedAt: ue.UpdatedAt,
+		UserPosts: ue.UserPosts,
+		Profile: UserProfileDTO{
+			Bio:         ue.Profile.Bio,
+			PictureURL:  ue.Profile.PictureURL,
+			SocialLinks: socialLinks,
+		},
+	}
 }
